PicPay-Pedro - Separando as responsabilidades dos construtores para que uma classe faça apenas uma coisa. - Separando os DTOs de transação em Input e Output - Separando os DTOs em dtos e externalDTOS - Criando um enum para sabermos qual foi o exato erro, porque desta forma não é necessário se amarrar a string de erro.

+++ src/main/java/com/PicpayTransferSystem/dtos/PersonDTO.java
@@ -1,5 +1,7 @@
 package com.PicpayTransferSystem.dtos;
 
+import java.math.BigDecimal;
+
 import lombok.Getter;
 import lombok.Setter;
 
@@ -9,5 +11,5 @@ public class PersonDTO {
     private String fullName;
     private String email; 
     private String document;
-    private double saldoinicial;
+    private BigDecimal saldoinicial;
 }
