+++ src/main/java/com/PicpayTransferSystem/interfaces/IAccountService.java
@@ -1,8 +1,10 @@
 package com.PicpayTransferSystem.interfaces;
 
+import java.math.BigDecimal;
+
 import com.PicpayTransferSystem.entities.PersonEntity;
 
 public interface IAccountService {
-    void createNewAccount(PersonEntity personEntity, double initialBalance);
+    void createNewAccount(PersonEntity personEntity, BigDecimal initialBalance);
 
 }
